core: split BillableCall\MigrateFromTrunksCdr service
